# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI-Enabled Government HR Recruitment & Workforce Management Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user and agency
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login and get JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # AGENCIES
...
  /candidates:
    get:
      operationId: getCandidates
      tags: [candidates]
      summary: List candidates
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of candidates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Candidate"
    post:
      operationId: createCandidate
      tags: [candidates]
      summary: Create a candidate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCandidateRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
  /candidates/{id}:
    get:
      operationId: getCandidate
      tags: [candidates]
      summary: Get candidate by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Candidate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateCandidate
      tags: [candidates]
      summary: Update candidate
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCandidateRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # APPLICATIONS
...
  /contracts:
    get:
      operationId: getContracts
      tags: [contracts]
      summary: List contracts
      security:
        - bearerAuth: []
      parameters:
        - name: employee_id
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of contracts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contract"
    post:
      operationId: createContract
      tags: [contracts]
      summary: Create a contract
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContractRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
  /contracts/{id}:
    get:
      operationId: getContract
      tags: [contracts]
      summary: Get contract by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Contract
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateContract
      tags: [contracts]
      summary: Update contract
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateContractRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # AI SCORES
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # Auth Schemas
    RegisterRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        agencyName:
          type: string
        agencyType:
          type: string
        roleName:
          type: string
      required:
        - name
        - email
        - password
        - agencyName
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
      required:
        - token
        - user
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        roleId:
          type: ["integer", "null"]
        agencyId:
          type: ["integer", "null"]
        status:
          type: string
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - status

    # Agency Schemas
...
    Candidate:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        cvUrl:
          type: ["string", "null"]
        parsedData:
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - email
    CreateCandidateRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        cvUrl:
          type: ["string", "null"]
      required:
        - name
        - email
    UpdateCandidateRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        cvUrl:
          type: ["string", "null"]
        parsedData:
          nullable: true

    # Application Schemas
...
    Contract:
      type: object
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        startDate:
          type: string
        endDate:
          type: ["string", "null"]
        type:
          type: string
        status:
          type: string
        documentUrl:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - employeeId
        - startDate
        - type
        - status
    CreateContractRequest:
      type: object
      properties:
        employeeId:
          type: integer
        startDate:
          type: string
        endDate:
          type: ["string", "null"]
        type:
          type: string
        documentUrl:
          type: ["string", "null"]
      required:
        - employeeId
        - startDate
    UpdateContractRequest:
      type: object
      properties:
        endDate:
          type: ["string", "null"]
        type:
          type: string
        status:
          type: string
        documentUrl:
          type: ["string", "null"]

    # AI Score Schemas
...